+++ src/main/scala/com/programmaticallyspeaking/ncd/chrome/domains/Debugger.scala
@@ -51,6 +51,11 @@ object Debugger extends Logging {
     */
   case class getPossibleBreakpoints(start: Location, end: Option[Location])
 
+  /**
+    * Continues execution until specific location is reached.
+    */
+  case class continueToLocation(location: Location)
+
   object ScriptParsedEventParams {
     def apply(script: Script): ScriptParsedEventParams = new ScriptParsedEventParams(script.id,
       script.url.toString,
@@ -157,6 +162,8 @@ class Debugger(filePublisher: FilePublisher, scriptHost: ScriptHost) extends Dom
   private val emittedScripts = mutable.Map[String, String]()
 
   private var lastCallFrameList: Option[Seq[CallFrame]] = None
+
+  private var temporaryBreakpointIds = Set[String]()
 
   private def emitScriptParsedEvent(script: Script) = {
     val hash = script.contentsHash()
@@ -296,10 +303,32 @@ class Debugger(filePublisher: FilePublisher, scriptHost: ScriptHost) extends Dom
 
     case Debugger.restartFrame(callFrameId) =>
       stackFramesToCallFrames(scriptHost.restartStackFrame(callFrameId))
+
+    case Debugger.continueToLocation(location) =>
+      // DevTools always use column 0 (there's a comment: "Always use 0 column."), but if we pass a column to
+      // the host, it will be too picky, so pass no column at all.
+      val scriptLocation = ScriptLocation(location.lineNumber + 1, None)
+      scriptHost.setBreakpoint(ScriptIdentity.fromId(location.scriptId), scriptLocation, None) match {
+        case Some(bp) =>
+          log.debug(s"Continue to location with temporary breakpoint ID ${bp.breakpointId}")
+          temporaryBreakpointIds += bp.breakpointId
+
+          scriptHost.resume()
+
+        case None =>
+          throw new IllegalArgumentException(s"Failed to continue to location $location, couldn't set a breakpoint there.")
+      }
   }
 
   override protected def handleScriptEvent: PartialFunction[ScriptEvent, Unit] = {
-    case hb: HitBreakpoint => pauseBasedOnBreakpoint(hb)
+    case hb: HitBreakpoint =>
+      pauseBasedOnBreakpoint(hb).foreach(breakpointId => {
+        if (temporaryBreakpointIds.contains(breakpointId)) {
+          // This was a temporary breakpoint, so remove it
+          temporaryBreakpointIds -= breakpointId
+          scriptHost.removeBreakpointById(breakpointId)
+        }
+      })
 
     case ScriptAdded(script) => self ! EmitScriptParsed(script)
 
@@ -326,7 +355,7 @@ class Debugger(filePublisher: FilePublisher, scriptHost: ScriptHost) extends Dom
     }
   }
 
-  private def pauseBasedOnBreakpoint(hitBreakpoint: HitBreakpoint): Unit = {
+  private def pauseBasedOnBreakpoint(hitBreakpoint: HitBreakpoint): Option[String] = {
     val callFrames = stackFramesToCallFrames(hitBreakpoint.stackFrames)
     hitBreakpoint.stackFrames.headOption match {
       case Some(sf) =>
@@ -335,9 +364,11 @@ class Debugger(filePublisher: FilePublisher, scriptHost: ScriptHost) extends Dom
         
         val params = PausedEventParams(callFrames, "other", Seq(sf.breakpoint.breakpointId))
         emitEvent("Debugger.paused", params)
+        Some(sf.breakpoint.breakpointId)
 
       case None =>
         log.warn("Unexpected! Got a HitBreakpoint without stack frames!")
+        None
     }
   }